fix: fix middleware 'sanctum' -> 'auth:santcum', added 'message' into service method - removed user return, switched st code and etc...

+++ routes/auth/auth.php
@@ -19,8 +19,9 @@ Route::prefix('auth')->group(function () {
 
         Route::get('tokens', [AuthController::class, 'tokens']);
 
-        Route::delete('tokens/{token}', [AuthController::class, 'deleteToken'])->whereNumber('token')->middleware('sanctum');
+        Route::delete('tokens/{token}', [AuthController::class, 'deleteToken'])->whereNumber('token')->middleware('auth:sanctum');
 
     });
 
 });
+
